+++ pdfalyzer/config.py
@@ -1,3 +1,4 @@
+from argparse import Namespace
 import importlib.resources
 import logging
 from os import environ, pardir, path
@@ -15,16 +16,19 @@ LOG_DIR_ENV_VAR = 'PDFALYZER_LOG_DIR'
 
 # 3rd part pdf-parser.py
 PDF_PARSER_EXECUTABLE_ENV_VAR = 'PDFALYZER_PDF_PARSER_PY_PATH'
+DEFAULT_PDF_PARSER_EXECUTABLE = path.join(PROJECT_ROOT, 'tools', 'pdf-parser.py')
 
 YaralyzerConfig.LOG_DIR = environ.get(LOG_DIR_ENV_VAR)
 YaralyzerConfig.LOG_LEVEL = logging.getLevelName(environ.get(LOG_LEVEL_ENV_VAR, 'WARN'))
-#set_log_level(YaralyzerConfig.LOG_LEVEL)
 
 
 class PdfalyzerConfig:
-    JAVSCRIPT_KEYWORD_ALERT_THRESHOLD = 2
-    QUOTE_TYPE = None
-    DEFAULT_PDF_PARSER_EXECUTABLE = path.join(PROJECT_ROOT, 'tools', 'pdf-parser.py')
+    _parsed_args: Namespace = Namespace()
+
+    @classmethod
+    def args(cls):
+        """Args parsed by argparse at startup."""
+        return cls._parsed_args
 
     # Path to Didier Stevens's pdf-parser.py
     if is_env_var_set_and_not_false(PDF_PARSER_EXECUTABLE_ENV_VAR):